requirements.txt
flask
faster-whisper
python-dotenv